Update FlxDrawQuadsItem.hx
I have an issue with html5 target, if I change the color of a FlxSprite, others FlxSprite sharing the same bitmap disappear.
After a few hours, I fixed the issue by removing this line "if (colored || hasColorOffsets)", the "hasColorTransform" on the shader seems to persist with a "true" value but the colorMultipliers and colorOffsets are set to 0. Maybe there's a better option.
In the addQuad function, colorMultipliers and colorOffset of the FlxDrawQuadsItem are correctly set by default, so in all cases we can set these values to the shader.